kline chart: move color math helpers to theme.rs
